Move file and adjust namespace.
