+++ osu.Game.Rulesets.Karaoke.Tests/Edit/Lyrics/CaretPosition/Algorithms/CreateTextTagCaretPositionAlgorithmTest.cs
@@ -3,12 +3,12 @@
 
 using System.Linq;
 using NUnit.Framework;
-using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Algorithms;
 using osu.Game.Rulesets.Karaoke.Edit.Lyrics.CaretPosition;
+using osu.Game.Rulesets.Karaoke.Edit.Lyrics.CaretPosition.Algorithms;
 using osu.Game.Rulesets.Karaoke.Objects;
 using osu.Game.Rulesets.Karaoke.Tests.Helper;
 
-namespace osu.Game.Rulesets.Karaoke.Tests.Edit.Lyrics.Algorithms
+namespace osu.Game.Rulesets.Karaoke.Tests.Edit.Lyrics.CaretPosition.Algorithms
 {
     [TestFixture]
     public class CreateTextTagCaretPositionAlgorithmTest : BaseCaretPositionAlgorithmTest<CreateTextTagCaretPositionAlgorithm, CreateTextTagCaretPosition>
