+++ osu.Game.Rulesets.Karaoke.Tests/Edit/Lyrics/CaretPosition/Algorithms/CuttingCaretPositionAlgorithmTest.cs
@@ -3,23 +3,26 @@
 
 using System.Linq;
 using NUnit.Framework;
-using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Algorithms;
 using osu.Game.Rulesets.Karaoke.Edit.Lyrics.CaretPosition;
+using osu.Game.Rulesets.Karaoke.Edit.Lyrics.CaretPosition.Algorithms;
 using osu.Game.Rulesets.Karaoke.Objects;
 
-namespace osu.Game.Rulesets.Karaoke.Tests.Edit.Lyrics.Algorithms
+namespace osu.Game.Rulesets.Karaoke.Tests.Edit.Lyrics.CaretPosition.Algorithms
 {
     [TestFixture]
-    public class TypingCaretPositionAlgorithmTest : BaseCaretPositionAlgorithmTest<TypingCaretPositionAlgorithm, TextCaretPosition>
+    public class CuttingCaretPositionAlgorithmTest : BaseCaretPositionAlgorithmTest<CuttingCaretPositionAlgorithm, TextCaretPosition>
     {
         protected const int INDEX_EXIST_TAG = -1;
 
-        [TestCase(nameof(singleLyric), 0, 0, true)]
-        [TestCase(nameof(singleLyric), 0, 4, true)]
+        [TestCase(nameof(singleLyric), 0, 1, true)]
+        [TestCase(nameof(singleLyric), 0, 3, true)]
+        [TestCase(nameof(singleLyric), 0, 0, false)]
+        [TestCase(nameof(singleLyric), 0, 4, false)]
         [TestCase(nameof(singleLyric), 0, 5, false)]
         [TestCase(nameof(singleLyric), 0, -1, false)]
-        [TestCase(nameof(singleLyricWithNoText), 0, 0, true)] // it's still movable even text is empty or null.
+        [TestCase(nameof(singleLyricWithNoText), 0, 0, true)] // It's not cuttable if no lyric text.
         [TestCase(nameof(singleLyricWithNoText), 0, 1, false)]
+        [TestCase(nameof(singleLyricWithNoText), 0, -1, false)]
         public void TestPositionMovable(string sourceName, int lyricIndex, int index, bool movable)
         {
             var lyrics = GetLyricsByMethodName(sourceName);
@@ -29,11 +32,10 @@ namespace osu.Game.Rulesets.Karaoke.Tests.Edit.Lyrics.Algorithms
             TestPositionMovable(lyrics, caretPosition, movable);
         }
 
-        [TestCase(nameof(singleLyric), 0, 0, NOT_EXIST, INDEX_EXIST_TAG)] // cannot move up if at top index.
-        [TestCase(nameof(singleLyricWithNoText), 0, 0, NOT_EXIST, INDEX_EXIST_TAG)]
-        [TestCase(nameof(twoLyricsWithText), 1, 0, 0, 0)]
-        [TestCase(nameof(threeLyricsWithSpacing), 2, 0, 1, 0)]
-        [TestCase(nameof(threeLyricsWithSpacing), 2, 3, 1, 0)]
+        [TestCase(nameof(singleLyric), 0, 1, NOT_EXIST, NOT_EXIST)] // cannot move up if at top index.
+        [TestCase(nameof(twoLyricsWithText), 1, 1, 0, 1)]
+        [TestCase(nameof(threeLyricsWithSpacing), 2, 1, 0, 1)]
+        [TestCase(nameof(threeLyricsWithSpacing), 2, 2, 0, 2)]
         public void TestMoveUp(string sourceName, int lyricIndex, int index, int newLyricIndex, int newIndex)
         {
             var lyrics = GetLyricsByMethodName(sourceName);
@@ -44,11 +46,11 @@ namespace osu.Game.Rulesets.Karaoke.Tests.Edit.Lyrics.Algorithms
             TestMoveUp(lyrics, caretPosition, newCaretPosition);
         }
 
-        [TestCase(nameof(singleLyric), 0, 0, NOT_EXIST, INDEX_EXIST_TAG)] // cannot move down if at bottom index.
-        [TestCase(nameof(singleLyricWithNoText), 0, 0, NOT_EXIST, INDEX_EXIST_TAG)]
-        [TestCase(nameof(twoLyricsWithText), 0, 0, 1, 0)]
-        [TestCase(nameof(threeLyricsWithSpacing), 0, 0, 1, 0)]
-        [TestCase(nameof(threeLyricsWithSpacing), 0, 4, 1, 0)]
+        [TestCase(nameof(singleLyric), 0, 1, NOT_EXIST, INDEX_EXIST_TAG)] // cannot move down if at bottom index.
+        [TestCase(nameof(twoLyricsWithText), 0, 1, 1, 1)]
+        [TestCase(nameof(twoLyricsWithText), 0, 3, 1, 2)]
+        [TestCase(nameof(threeLyricsWithSpacing), 0, 1, 2, 1)]
+        [TestCase(nameof(threeLyricsWithSpacing), 0, 3, 2, 2)]
         public void TestMoveDown(string sourceName, int lyricIndex, int index, int newLyricIndex, int newIndex)
         {
             var lyrics = GetLyricsByMethodName(sourceName);
@@ -59,10 +61,9 @@ namespace osu.Game.Rulesets.Karaoke.Tests.Edit.Lyrics.Algorithms
             TestMoveDown(lyrics, caretPosition, newCaretPosition);
         }
 
-        [TestCase(nameof(singleLyric), 0, 0, NOT_EXIST, INDEX_EXIST_TAG)]
-        [TestCase(nameof(singleLyricWithNoText), 0, 0, NOT_EXIST, INDEX_EXIST_TAG)]
-        [TestCase(nameof(twoLyricsWithText), 1, 0, 0, 4)]
-        [TestCase(nameof(threeLyricsWithSpacing), 2, 0, 1, 0)]
+        [TestCase(nameof(singleLyric), 0, 1, NOT_EXIST, INDEX_EXIST_TAG)]
+        [TestCase(nameof(twoLyricsWithText), 1, 1, 0, 3)]
+        [TestCase(nameof(threeLyricsWithSpacing), 2, 1, 0, 3)]
         [TestCase(nameof(threeLyricsWithSpacing), 2, 3, 2, 2)]
         public void TestMoveLeft(string sourceName, int lyricIndex, int index, int newLyricIndex, int newIndex)
         {
@@ -74,11 +75,10 @@ namespace osu.Game.Rulesets.Karaoke.Tests.Edit.Lyrics.Algorithms
             TestMoveLeft(lyrics, caretPosition, newCaretPosition);
         }
 
-        [TestCase(nameof(singleLyric), 0, 4, NOT_EXIST, INDEX_EXIST_TAG)]
-        [TestCase(nameof(singleLyricWithNoText), 0, 0, NOT_EXIST, INDEX_EXIST_TAG)]
-        [TestCase(nameof(twoLyricsWithText), 0, 4, 1, 0)]
-        [TestCase(nameof(threeLyricsWithSpacing), 0, 4, 1, 0)]
-        [TestCase(nameof(threeLyricsWithSpacing), 0, 3, 0, 4)]
+        [TestCase(nameof(singleLyric), 0, 3, NOT_EXIST, INDEX_EXIST_TAG)]
+        [TestCase(nameof(twoLyricsWithText), 0, 3, 1, 1)]
+        [TestCase(nameof(threeLyricsWithSpacing), 0, 3, 2, 1)]
+        [TestCase(nameof(threeLyricsWithSpacing), 0, 2, 0, 3)]
         public void TestMoveRight(string sourceName, int lyricIndex, int index, int newLyricIndex, int newIndex)
         {
             var lyrics = GetLyricsByMethodName(sourceName);
@@ -89,10 +89,10 @@ namespace osu.Game.Rulesets.Karaoke.Tests.Edit.Lyrics.Algorithms
             TestMoveRight(lyrics, caretPosition, newCaretPosition);
         }
 
-        [TestCase(nameof(singleLyric), 0, 0)]
-        [TestCase(nameof(singleLyricWithNoText), 0, 0)]
-        [TestCase(nameof(twoLyricsWithText), 0, 0)]
-        [TestCase(nameof(threeLyricsWithSpacing), 0, 0)]
+        [TestCase(nameof(singleLyric), 0, 1)]
+        [TestCase(nameof(singleLyricWithNoText), NOT_EXIST, INDEX_EXIST_TAG)]
+        [TestCase(nameof(twoLyricsWithText), 0, 1)]
+        [TestCase(nameof(threeLyricsWithSpacing), 0, 1)]
         public void TestMoveToFirst(string sourceName, int lyricIndex, int index)
         {
             var lyrics = GetLyricsByMethodName(sourceName);
@@ -102,10 +102,10 @@ namespace osu.Game.Rulesets.Karaoke.Tests.Edit.Lyrics.Algorithms
             TestMoveToFirst(lyrics, caretPosition);
         }
 
-        [TestCase(nameof(singleLyric), 0, 4)]
-        [TestCase(nameof(singleLyricWithNoText), 0, 0)]
-        [TestCase(nameof(twoLyricsWithText), 1, 3)]
-        [TestCase(nameof(threeLyricsWithSpacing), 2, 3)]
+        [TestCase(nameof(singleLyric), 0, 3)]
+        [TestCase(nameof(singleLyricWithNoText), NOT_EXIST, INDEX_EXIST_TAG)]
+        [TestCase(nameof(twoLyricsWithText), 1, 2)]
+        [TestCase(nameof(threeLyricsWithSpacing), 2, 2)]
         public void TestMoveToLast(string sourceName, int lyricIndex, int index)
         {
             var lyrics = GetLyricsByMethodName(sourceName);
@@ -115,13 +115,13 @@ namespace osu.Game.Rulesets.Karaoke.Tests.Edit.Lyrics.Algorithms
             TestMoveToLast(lyrics, caretPosition);
         }
 
-        [TestCase(nameof(singleLyric), 0, 0)]
-        [TestCase(nameof(singleLyricWithNoText), 0, 0)]
-        public void TestMoveToTarget(string sourceName, int lyricIndex, int index)
+        [TestCase(nameof(singleLyric), 0)]
+        [TestCase(nameof(singleLyricWithNoText), 0)]
+        public void TestMoveToTarget(string sourceName, int lyricIndex)
         {
             var lyrics = GetLyricsByMethodName(sourceName);
             var lyric = lyrics[lyricIndex];
-            var caretPosition = CreateTextCaretPosition(lyrics, lyricIndex, index);
+            var caretPosition = CreateTextCaretPosition(lyrics, lyricIndex, 1);
 
             // Check move to target position.
             TestMoveToTarget(lyrics, lyric, caretPosition);
