+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/CaretPosition/Algorithms/TextTagCaretPositionAlgorithm.cs
@@ -3,10 +3,9 @@
 
 using System;
 using System.Linq;
-using osu.Game.Rulesets.Karaoke.Edit.Lyrics.CaretPosition;
 using osu.Game.Rulesets.Karaoke.Objects;
 
-namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Algorithms
+namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.CaretPosition.Algorithms
 {
     public abstract class TextTagCaretPositionAlgorithm<T> : CaretPositionAlgorithm<T> where T : class, ITextTagCaretPosition
     {
